% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_setup_metadata}
\alias{hydb_setup_metadata}
\title{Setup Metadata}
\usage{
hydb_setup_metadata(point, column_map = NULL)
}
\arguments{
\item{point}{An sf POINT cleaned.}

\item{column_map}{A vector of names to change in point sf column names, (NULL) default.}
}
\value{
A \code{tibble()} ready for appending to \code{hydb}.
}
\description{
Setup Metadata
}
\section{Functions}{
\itemize{
\item \code{hydb_setup_metadata()}: function to setup user submitted stations metadata.

}}
