% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{convert_dates}
\alias{convert_dates}
\title{Convert to SQL dates}
\usage{
convert_dates(condition)
}
\arguments{
\item{condition}{A list}
}
\value{
A list with updated date syntax for SQL.
}
\description{
Convert to SQL dates
}
