% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{find_closest_match}
\alias{find_closest_match}
\title{Find Closest Match}
\usage{
find_closest_match(text, keys)
}
\arguments{
\item{text}{A character string}

\item{keys}{A character string}
}
\value{
A character string
}
\description{
Find Closest Match
}
