% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{ftbl}
\alias{ftbl}
\title{Fetch Table}
\usage{
ftbl(table, mydb, tbl_only, ...)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{mydb}{A connected database object.}

\item{tbl_only}{A logical indicating whether to \link[dplyr]{collect} the table.}

\item{...}{Option to pass calls to \code{dplyr::filter()}.}
}
\value{
A table
}
\description{
Fetch Table
}
