% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{graphingMod}
\alias{graphingMod}
\title{Shiny Module Server for DT uploading}
\usage{
graphingMod(input, output, session)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{values}{A reactive Values list to pass}

\item{sheet}{A reactive Values list to pass}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for DT uploading
}
