% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{hydb_append_table}
\alias{hydb_append_table}
\title{Append DB Table}
\usage{
hydb_append_table(
  data,
  table = NULL,
  check = TRUE,
  copy_path = TRUE,
  path_to_copy = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame to append.}

\item{table}{character. DB table to append to.}

\item{check}{logical. Checking details before appending.}

\item{copy_path}{logical. Whether to make a copy of db.}

\item{path_to_copy}{character. Where to write the copy of db if \code{copy = TRUE}.}

\item{...}{Arguments to pass to \code{hydb_connect()}.}
}
\value{
Nothing. Side effect to \code{hydb}.
}
\description{
Append DB Table
}
