% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestnuances.R
\name{hydb_bulk_excel}
\alias{hydb_bulk_excel}
\title{Bulk load Excel Sheets}
\usage{
hydb_bulk_excel(path, skip = 0)
}
\arguments{
\item{path}{A character path to .xlsx file.}

\item{skip}{A numeric.}
}
\value{
A tibble.
}
\description{
Bulk load Excel Sheets
}
\note{
Skips the first sheet. Will make more dynamic to handle
}
