% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{hydb_clean_duplicate_metadata}
\alias{hydb_clean_duplicate_metadata}
\title{Clean Spatial Duplicates}
\usage{
hydb_clean_duplicate_metadata(point)
}
\arguments{
\item{point}{An sf POINT object}
}
\value{
An sf object
}
\description{
Get cleaned duplicates.
}
