% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydb_connect}
\alias{hydb_connect}
\title{Connect to hydb}
\usage{
hydb_connect()
}
\value{
Nothing. Side effect connecting to hydb.
}
\description{
Connect to hydb
}
\note{
Must be a memeber of the \verb{USDA Northern Region Hydrology} sharepoint group.
}
