% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_conversions}
\alias{hydb_conversions}
\title{Conversions}
\usage{
hydb_conversions(x, type)
}
\arguments{
\item{x}{data.frame}

\item{type}{character.}
}
\value{
Side effect transformation on data.frame
}
\description{
Convert Metrics
}
\note{
When using 'c' for Celsius, this converts Fahrenheit to Celsius and vice versa for 'f'.
}
