% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydb_copy_db}
\alias{hydb_copy_db}
\title{Copy hydb.sqlite and keep only the latest backup.}
\usage{
hydb_copy_db(
  path =
    "C:/Users/joshualerickson/USDA/Northern Region Hydrology - Documents/data-madness/hydb"
)
}
\arguments{
\item{path}{Directory where hydb.sqlite and backups are stored.}
}
\value{
TRUE if copy succeeded and cleanup was performed.
}
\description{
Copy hydb.sqlite and keep only the latest backup.
}
