% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_daily_transform}
\alias{hydb_daily_transform}
\title{Transform IV to DV}
\usage{
hydb_daily_transform(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A \code{tibble()}.
}
\description{
Transform IV to DV
}
\note{
Need to have columns \code{dt}, \code{sid} and \verb{iv_*} and \verb{iv_*} needs to be column position 2.
}
