% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydb.R
\name{hydb_disconnect}
\alias{hydb_disconnect}
\title{Disconnect to hydb}
\usage{
hydb_disconnect(db = NULL)
}
\value{
Nothing. Side effect connecting to hydb.
}
\description{
Disconnect to hydb
}
\note{
Must be a member of the \verb{USDA Northern Region Hydrology} sharepoint group.
}
