% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{hydb_fields}
\alias{hydb_fields}
\title{hydb Table List}
\usage{
hydb_fields(table = table, network = NULL)
}
\arguments{
\item{table}{A character.}

\item{network}{NULL. Used in shiny application.}
}
\value{
A list of tables in the hydb database.
}
\description{
hydb Table List
}
