% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{hydb_handling_duplicates}
\alias{hydb_handling_duplicates}
\title{Handling Duplicates}
\usage{
hydb_handling_duplicates(data, existing_data)
}
\arguments{
\item{data}{Data selected in GUI.}

\item{existing_data}{Data with same sid in hydb.}
}
\value{
Nothing. Side-effect to database.
}
\description{
Handling Duplicates
}
