% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_qaqc}
\alias{hydb_qaqc}
\title{QA/QC table}
\usage{
hydb_qaqc(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A \code{tibble()}.
}
\description{
QA/QC table
}
\note{
Need to be a table from \code{hydb}.
}
