% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydb_setup_metadata}
\alias{hydb_setup_metadata}
\title{Setup Metadata}
\usage{
hydb_setup_metadata(point)
}
\arguments{
\item{point}{An sf POINT cleaned.}
}
\value{
A \code{tibble()} ready for appending to hydb.
}
\description{
Setup Metadata
}
