% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_station_info}
\alias{hydb_station_info}
\title{Stations in Tables}
\usage{
hydb_station_info(data, hydbtables = "all")
}
\arguments{
\item{data}{data.frame with station sid's or character vector of sid's.}

\item{hydbtables}{character. tables to select from, default 'all'.}
}
\value{
Nothing. Side effect to \code{hydb}.
}
\description{
Stations in Tables
}
