% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{hydbModUI}
\alias{hydbModUI}
\title{Shiny Module UI for DT uploading}
\usage{
hydbModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
UI function for Shiny module
}
\description{
A shiny Module to.
}
