% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{longest_streak}
\alias{longest_streak}
\title{Longest Streak}
\usage{
longest_streak(num_list)
}
\arguments{
\item{num_list}{A list with numeric input.}
}
\value{
A list
}
\description{
Longest Streak
}
