% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{slice_hydb}
\alias{slice_hydb}
\title{Slice for tbl_sql}
\usage{
slice_hydb(.data, ...)
}
\arguments{
\item{.data}{tbl_sql class}

\item{...}{dots to pass to \code{dplyr} verbs}
}
\description{
Slice for tbl_sql
}
