% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_prep_data}
\alias{hydb_prep_data}
\title{Prep 'hydb' Table}
\usage{
hydb_prep_data(data, wy_month = 10)
}
\arguments{
\item{data}{A \code{hydb} table.}

\item{wy_month}{Numeric. Month starting water year.}
}
\value{
\code{data} but with added attributes
}
\description{
Prep 'hydb' Table
}
\section{Functions}{
\itemize{
\item \code{hydb_prep_data()}: function preps the data for more analysis.

}}
