% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_hydb}
\alias{update_hydb}
\title{Fetch Data from Hydrological Database.}
\usage{
update_hydb(table, new_values, conditions, network = NULL)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{new_values}{A list of column and value to change for SQL.}

\item{conditions}{A list of conditions to be met for SQL.}

\item{network}{NULL. Used in shiny application.}
}
\value{
Nothing.
}
\description{
Fetch Data from Hydrological Database.
}
\note{
A list
}
