% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{uptbl}
\alias{uptbl}
\title{Update Table}
\usage{
uptbl(table, new_values, conditions, mydb)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{new_values}{A list of column and value to change for SQL.}

\item{conditions}{A list of conditions to be met for SQL.}

\item{mydb}{A database object}
}
\value{
Nothing. This function updates the database
}
\description{
Update Table
}
