% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemanip.R
\name{hydb_setup_metadata}
\alias{hydb_setup_metadata}
\title{Setup Metadata}
\usage{
hydb_setup_metadata(point)
}
\arguments{
\item{point}{An sf POINT cleaned.}
}
\value{
A \code{tibble()} ready for appending to \code{hydb}.
}
\description{
Setup Metadata
}
\section{Functions}{
\itemize{
\item \code{hydb_setup_metadata()}: function to setup user submitted stations metadata.

}}
