% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_hydb}
\alias{fetch_hydb}
\title{Fetch Data from Hydrological Database.}
\usage{
fetch_hydb(table, sid = NULL, network = TRUE)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{sid}{A character vector of station ids. NULL (default).}

\item{network}{logical. Whether to connect via the network or server. TRUE (default).}
}
\value{
A data.frame
}
\description{
Fetch Data from Hydrological Database.
}
