% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_hydb}
\alias{fetch_hydb}
\title{Fetch Data from Hydrological Database.}
\usage{
fetch_hydb(table, sid = NULL, network = NULL)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{sid}{A character vector of station ids. NULL (default).}

\item{network}{NULL. Used in shiny application.}
}
\value{
A data.frame
}
\description{
Fetch Data from Hydrological Database.
}
