% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{ftbl}
\alias{ftbl}
\title{Fetch Table}
\usage{
ftbl(table, sid, mydb, tbl_only)
}
\arguments{
\item{table}{A character vector of table names, e.g. 'station_metadata', 'flow_dv', etc.}

\item{sid}{A character vector of station ids. NULL (default).}

\item{mydb}{A connected database object.}

\item{tbl_only}{A logical indicating whether to \link[dplyr]{collect} the table.}
}
\value{
A table
}
\description{
Fetch Table
}
