% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydb.R
\name{hydb_connect}
\alias{hydb_connect}
\title{Connect to hydb}
\usage{
hydb_connect(path = NULL)
}
\arguments{
\item{path}{A character vector file path to \code{.sqlite} database.}
}
\value{
Nothing. Side effect connecting to hydb.
}
\description{
Connect to hydb
}
\note{
Must be a member of the \verb{USDA Northern Region Hydrology} sharepoint group if \code{path = NULL}.
}
