% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{hydb_tables}
\alias{hydb_tables}
\title{hydb Table List}
\usage{
hydb_tables(network = NULL, ...)
}
\arguments{
\item{network}{NULL. Used in shiny application.}

\item{...}{Arguments to pass to \code{hydb_connect()}}
}
\value{
A list of tables in the hydb database.
}
\description{
hydb Table List
}
